requirements.txt
pandas
pyupbit
python-dotenv
python-telegram-bot
